+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GoodTiming">
         <activity
-            android:name=".Todomain"
+            android:name=".Utils.Todomain"
             android:exported="false" />
         <activity
-            android:name=".WalkthroughActivity"
+            android:name=".Utils.WalkthroughActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".Utils.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
             <intent-filter>
