+++ app/src/main/java/edu/fsu/cs/goodtiming/MainActivity.java
@@ -9,6 +9,9 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import edu.fsu.cs.goodtiming.Utils.CalendarFragment;
+import edu.fsu.cs.goodtiming.Utils.SessionFragment;
+
 public class MainActivity extends AppCompatActivity implements
         EventFragment.OnEventFragmentInteractionListener,
         SessionFragment.OnSessionFragmentInteractionListener,
