+++ app/src/main/java/edu/fsu/cs/goodtiming/Utils/NewEventFragment.java
@@ -1,12 +1,10 @@
-package edu.fsu.cs.goodtiming;
+package edu.fsu.cs.goodtiming.Utils;
 
 import static androidx.constraintlayout.helper.widget.MotionEffect.TAG;
 
 import android.app.DatePickerDialog;
-import android.app.TimePickerDialog;
 import android.os.Bundle;
 
-import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
@@ -17,12 +15,12 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.EditText;
-import android.widget.TimePicker;
 
 import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
-import java.util.Date;
+
+import edu.fsu.cs.goodtiming.EventFragment;
+import edu.fsu.cs.goodtiming.R;
 
 
 public class NewEventFragment extends Fragment implements DatePickerDialog.OnDateSetListener {
