+++ app/src/main/java/edu/fsu/cs/goodtiming/Utils/NewTaskFragment.java
@@ -1,4 +1,4 @@
-package edu.fsu.cs.goodtiming;
+package edu.fsu.cs.goodtiming.Utils;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import edu.fsu.cs.goodtiming.R;
 
 
 public class NewTaskFragment extends Fragment {
