+++ app/src/main/java/edu/fsu/cs/goodtiming/Utils/OnDialogCloseListner.java
@@ -1,4 +1,4 @@
-package edu.fsu.cs.goodtiming;
+package edu.fsu.cs.goodtiming.Utils;
 
 import android.content.DialogInterface;
