+++ app/src/main/java/edu/fsu/cs/goodtiming/Utils/SplashActivity.java
@@ -1,10 +1,12 @@
-package edu.fsu.cs.goodtiming;
+package edu.fsu.cs.goodtiming.Utils;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
+
+import edu.fsu.cs.goodtiming.R;
 
 public class SplashActivity extends AppCompatActivity {
