+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Splashprimarydark"
-    tools:context=".SplashActivity">
+    tools:context=".Utils.SplashActivity">
 
     <TextView
         android:id="@+id/tvTime"
