+++ app/src/main/res/layout/activity_todomain.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/main"
-    tools:context=".Todomain">
+    tools:context=".Utils.Todomain">
 
     <TextView
         android:layout_width="wrap_content"
