+++ app/src/main/res/layout/activity_walkthrough.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Splashprimarydark"
-    tools:context=".WalkthroughActivity">
+    tools:context=".Utils.WalkthroughActivity">
 
     <Button
         android:id="@+id/skipButton"
